# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Activepieces Documentation",
    "version": "0.3.6"
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
    "/v1/collections/{collectionId}": {
      "delete": {
        "summary": "Delete a collection",
        "tags": [
          "collection"
        ],
        "description": "Delete a collection",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "collectionId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get a collection",
        "tags": [
          "collection"
        ],
        "description": "Get a collection",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "collectionId",
            "required": true
          }
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "example": "9m7dpndtY3IUvUvCv5WY0",
                      "type": "string"
                    },
                    "created": {
                      "example": "2023-03-09T18:29:49.146Z",
                      "type": "string"
                    },
                    "update": {
                      "example": "2023-03-09T18:29:49.146Z",
                      "type": "string"
                    },
                    "projectId": {
                      "example": "9m7dpndtY3IUvUvCv5WY0",
                      "type": "string"
                    },
                    "displayName": {
                      "example": "Collection",
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "created",
                    "update",
                    "projectId",
                    "displayName"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Update a collection",
        "tags": [
          "collection"
        ],
        "description": "Update a collection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": "string"
                  }
                },
                "required": [
                  "displayName"
                ]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "collectionId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "example": "9m7dpndtY3IUvUvCv5WY0",
                      "type": "string"
                    },
                    "created": {
                      "example": "2023-03-09T18:29:49.146Z",
                      "type": "string"
                    },
                    "update": {
                      "example": "2023-03-09T18:29:49.146Z",
                      "type": "string"
                    },
                    "projectId": {
                      "example": "9m7dpndtY3IUvUvCv5WY0",
                      "type": "string"
                    },
                    "displayName": {
                      "example": "Collection",
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "created",
                    "update",
                    "projectId",
                    "displayName"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/collections/": {
      "get": {
        "summary": "List Collections",
        "tags": [
          "collection"
        ],
        "description": "List Collections",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "next": {
                      "example": "next-cursor",
                      "type": "string"
                    },
                    "previous": {
                      "example": "previous-cursor",
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "example": "9m7dpndtY3IUvUvCv5WY0",
                            "type": "string"
                          },
                          "created": {
                            "example": "2023-03-09T18:29:49.146Z",
                            "type": "string"
                          },
                          "update": {
                            "example": "2023-03-09T18:29:49.146Z",
                            "type": "string"
                          },
                          "projectId": {
                            "example": "9m7dpndtY3IUvUvCv5WY0",
                            "type": "string"
                          },
                          "displayName": {
                            "example": "Collection",
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "created",
                          "update",
                          "projectId",
                          "displayName"
                        ]
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a collection",
        "tags": [
          "collection"
        ],
        "description": "Create a collection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": "string"
                  }
                },
                "required": [
                  "displayName"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "example": "9m7dpndtY3IUvUvCv5WY0",
                      "type": "string"
                    },
                    "created": {
                      "example": "2023-03-09T18:29:49.146Z",
                      "type": "string"
                    },
                    "update": {
                      "example": "2023-03-09T18:29:49.146Z",
                      "type": "string"
                    },
                    "projectId": {
                      "example": "9m7dpndtY3IUvUvCv5WY0",
                      "type": "string"
                    },
                    "displayName": {
                      "example": "Collection",
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "created",
                    "update",
                    "projectId",
                    "displayName"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Self hosted"
    }
  ],
...
}